refactor: HistoryMenu component refactored
